+++ language/benchmarks/Cargo.toml
@@ -24,6 +24,7 @@ move-core-types = { path = "../move-core/types", version = "0.1.0" }
 move-lang = { path = "../move-lang", version = "0.0.1" }
 move-vm-runtime = { path = "../move-vm/runtime", version = "0.1.0" }
 move-vm-state = { path = "../move-vm/state", version = "0.1.0" }
+move-vm-types = { path = "../move-vm/types", version = "0.1.0" }
 vm = { path = "../vm", version = "0.1.0" }
 libra-vm = { path = "../libra-vm", version = "0.1.0" }
