+++ language/move-vm/types/Cargo.toml
@@ -11,6 +11,8 @@ edition = "2018"
 
 [dependencies]
 bit-vec = "0.6.1"
+mirai-annotations = "1.7.0"
+once_cell = "1.3.1"
 proptest = { version = "0.9.6", optional = true }
 sha2 = "0.8.0"
 serde = { version = "1.0", features = ["derive", "rc"] }
