+++ language/vm/src/lib.rs
@@ -15,7 +15,6 @@ pub mod errors;
 pub mod deserializer;
 pub mod file_format;
 pub mod file_format_common;
-pub mod gas_schedule;
 pub mod internals;
 #[cfg(any(test, feature = "fuzzing"))]
 pub mod proptest_types;
